Remove creation of default artifact for OneJar plugin, fixes #3
Since the OneJar type is derived from org.gradle.api.tasks.bundling.Jar all
necessary prerequisites are already there and the task doesn't need to create
and add a PublishArtifact itself.

The output of the OneJar task can (similar to other tasks of the type Jar) be
added via the 'artifacts' definition in the build file (see
http://www.gradle.org/docs/current/userguide/artifact_management.html). This
way the task implementing OneJar will automatically become a dependency of the
publish task (which in turn can publish the artifacts to a repository).